Add log4j2 feature
Added log4j2 dependency in pom.xml.
Created a log4j2 class with seperate methods for calling info, warn, error & debug that can be called in other classes. For now we wont save log messages in a file. But if we want to activate it, we can just uncomment the lines in log4j2.xml under the resources folder.

Rebased main to feature branch