Prepare SaveFile for testing
To be able to specify a path to a test folder during test running
a new method will be added called setDir(Path). Default folder will
still be used in user home folder if not specifying a Path with setDir.

Also removes \n from writeString call that otherwise creates an empty
line at the end.